Add info for clang without -integrated-as support
The -fblocks option can still be used with versions of clang that
do not have integrated assembler support.  The source file must
first be compiled to a .s file using -fblocks and then that file
compiled without -fblocks to the desired end product.

Add support for this (and a warning message) to checktests and
the README.txt file.