support shared library builds
Although use of the static library is recommended, support building
a shared library as well with the optional `-shared` option.

By default install both if they exist, but allow a single `-shared`
or `-static` to be passed to the installlib script to limit the
installation to just one or the other.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>