buildlib: supply -fPIC or -fpic if supported
Although this is the "blocks" runtime support library,
it does not necessarily need to be built by a compiler
that supports the compilation of blocks.

It's possible that the runtime support library is being
built by a compiler that does not support or understand
either the -fPIC or the -fpic option.

For this reason be cautious and only add the option
(preferring -fPIC) if the compiler actually groks it.

Suggested-by: Ron Olson <rolson@lbpc.com>
Signed-off-by: Kyle J. McKay <mackyle@gmail.com>